visualizer/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
chromadb>=1.0.0
numpy>=1.26.0
scikit-learn>=1.5.0
umap-learn>=0.5.0
pydantic>=2.0.0
